% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_convert_unit.R
\name{ct_convert_unit}
\alias{ct_convert_unit}
\title{Convert values between different units}
\usage{
ct_convert_unit(x, from, to, show_units = FALSE)
}
\arguments{
\item{x}{Numeric. Value(s) to convert.}

\item{from}{Character. The unit to convert from. Can be any synonym
e.g. \code{"meter"}, \code{"m"}, \code{"metre"}, \code{"feet"}, "\eqn{^\circ}", "um" (i.e \eqn{\mu m}), etc.}

\item{to}{Character. The unit to convert to. Can also be any synonym.}

\item{show_units}{Logical. If \code{TRUE}, returns the full table of supported units
and their synonyms instead of performing a conversion.}
}
\value{
A numeric vector of converted values.
}
\description{
Convert a numeric value from one unit to another.
It supports \strong{area}, \strong{distance (length)}, and \strong{angle} units.
}
\examples{
# Distance
ct_convert_unit(1000, "m", "km")
ct_convert_unit(1, "mile", "m")
ct_convert_unit(12, "inches", "ft")

# Area
ct_convert_unit(1, "acre", "m2")
ct_convert_unit(2, "km2", "hectare")

# Angle
ct_convert_unit(180, "deg", "rad")
ct_convert_unit(pi, "rad", "deg")

}
\seealso{
\code{\link[=units_table]{units_table()}} for supported units and synonyms.
}
