% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_create_hs.R
\name{ct_create_hs}
\alias{ct_create_hs}
\title{Create or add hierarchical subject values in image metadata}
\usage{
ct_create_hs(path, value = c(), intern = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{path}{A character vector specifying the full path of the image file.}

\item{value}{named character vector specifying the new hierarchical subjects to add.
Each value must have a parent specified as the name, e.g c("Species" = "Vulture").}

\item{intern}{TRUE if output should be returned as a character vector.}

\item{quiet}{Suppress output of the command itself.}

\item{...}{additional arguments to be passed to \code{\link[=system2]{system2()}}}
}
\description{
Create or add hierarchical subject values in image metadata
}
\examples{

# Image path
image_path <- file.path(system.file("img", package = "ct"), "large.jpeg")

# Get Hierarchical Subject from the image - Before use ct_create_hs()
ct_get_hs(path = image_path) #==> NULL

ct_create_hs(path = image_path, value = c("Species" = "Vulture"))

# Get Hierarchical Subject from the image - Before use ct_create_hs()
ct_get_hs(path = image_path) #==> "Species|Vulture"

}
