% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp.R
\name{ct_dp_example}
\alias{ct_dp_example}
\title{Read the Camtrap DP example dataset}
\usage{
ct_dp_example()
}
\value{
Camera Trap Data Package object.
}
\description{
Reads the \href{https://camtrap-dp.tdwg.org/example/}{Camtrap DP example dataset}.
This dataset is maintained and versioned with the Camtrap DP standard.
}
\concept{sample data}
