% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp.R
\name{ct_dp_read}
\alias{ct_dp_read}
\title{Read camera trap data package}
\usage{
ct_dp_read(file)
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}
}
\value{
A Camera Trap Data Package object.
}
\description{
Reads \href{https://camtrap-dp.tdwg.org}{Camera Trap Data Package} (Camtrap DP)
dataset into memory.
}
\section{Taxonomic information}{



Camtrap DP metadata has a \code{taxonomic} property that can contain extra
information for each \code{scientificName} found in observations.
Such information can include higher taxonomy (\code{family}, \code{order}, etc.) and
vernacular names in multiple languages.

The \code{read_camtrapdp()} function \strong{will automatically include this taxonomic
information in observations}, as extra columns starting with \code{taxon.}.
It will then update the \code{taxonomic} scope in the metadata to the unique
\code{\link[camtrapdp:taxa]{taxa()}} found in the data.

}

\section{Events}{



Observations can contain classifications at two levels:
\itemize{
\item \strong{Media-based} observations (\code{observationLevel = "media"}) are based on a
single media file and are directly linked to it via \code{mediaID}.
\item \strong{Event-based} observations (\code{observationLevel = "event"}) are based on an
event, defined as a combination of \code{eventID}, \code{eventStart} and \code{eventEnd}.
This event can consist of one or more media files, but is not directly linked
to these.
}

The \code{read_camtrapdp()} function \strong{will automatically assign \code{eventID}s to
media}, using \code{media.deploymentID = observations.deploymentID} and
\verb{observations.eventStart <= media.timestamp <= observations.eventEnd}.
Note that this can result in media being linked to multiple events (and thus
being duplicated), for example when events and sub-events were defined.

}

\examples{
\donttest{
file <- "https://raw.githubusercontent.com/tdwg/camtrap-dp/1.0/example/datapackage.json"
dp <- ct_dp_read(file)
}
}
