% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp.R
\name{ct_dp_table}
\alias{ct_dp_table}
\title{Get core tables}
\usage{
ct_dp_table(
  package,
  table = c("observations", "deployments", "media", "events", "taxa")
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
ct_read_dp().}

\item{table}{Character indicating the table to read - one "observations",
"deployments", or "media"}
}
\value{
A tibble of table specified
}
\description{
Acess table like observations, deployement, and media from data package.
}
\examples{
\donttest{
dp <- ct_dp_example()
ct_dp_table(dp, "deployments")
}

}
