% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp.R
\name{ct_dp_version}
\alias{ct_dp_version}
\title{Get Camtrap DP version
Extracts the version number used by a Camera Trap Data Package object.
This version number indicates what version of the \href{https://camtrap-dp.tdwg.org}{Camtrap DP standard} was used.}
\usage{
ct_dp_version(package)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
ct_read_dp().}
}
\value{
Camera Trap Data Package object.
}
\description{
Get Camtrap DP version
Extracts the version number used by a Camera Trap Data Package object.
This version number indicates what version of the \href{https://camtrap-dp.tdwg.org}{Camtrap DP standard} was used.
}
\details{
The version number is derived as follows:
\enumerate{
\item The \code{version} attribute, if defined.
\item A version number contained in \code{x$profile}, which is expected to
contain the URL to the used Camtrap DP standard.
\item \code{x$profile} in its entirety (can be \code{NULL}).
}
}
\examples{
dp <- ct_dp_example()
ct_dp_version(dp)

}
