% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_fit_ise.R
\name{ct_fit_ise}
\alias{ct_fit_ise}
\title{Estimate abundance from Instantaneous Sampling (ISE) Data}
\usage{
ct_fit_ise(
  data,
  deployment_data,
  sampling_frequency,
  sampling_length,
  study_area,
  study_start = NULL,
  study_end = NULL,
  quiet = FALSE
)
}
\arguments{
\item{data}{A tibble of camera trap detections. Must contain columns \code{cam}, \code{datetime}, and \code{count}.}

\item{deployment_data}{A tibble of camera deployments. Must contain columns \code{cam}, \code{start}, \code{end}, and \code{area}.}

\item{sampling_frequency}{Numeric. The number of seconds between the start of each sampling occasion.}

\item{sampling_length}{Numeric. The number of seconds to sample at each sampling occasion.}

\item{study_area}{Numeric. The size of the total study area in the same units as the camera viewshed area.}

\item{study_start}{POSIXct. The start of the study. Defaults to the minimum start time in \code{deployment_data}.}

\item{study_end}{POSIXct. The end of the study. Defaults to the maximum end time in \code{deployment_data}.}

\item{quiet}{Logical. Suppress status messages? Defaults to FALSE.}
}
\value{
A data.frame with the estimated abundance (\code{N}), its standard error (\code{SE}), and confidence intervals.
}
\description{
Estimate abundance from camera trap data using Instantaneous Sampling / point counts.
}
\section{References}{

Moeller, A. K. and P. M. Lukacs. 2021. spaceNtime: an R package for
estimating abundance
of unmarked animals using camera-trap photographs. Mammalian Biology.
\doi{10.1007/s42991-021-00181-8}

Moeller, A. K., P. M. Lukacs, and J. Horne. 2018. Three novel methods
to estimate abundance of unmarked animals using remote cameras. Ecosphere 9(8):
e02331. \doi{10.1002/ecs2.2331}
}

\examples{
data <- dplyr::tibble(
  cam = c(1, 1, 2, 2, 2),
  datetime = as.POSIXct(
    c(
      "2026-01-02 12:00:00",
      "2026-01-03 13:12:00",
      "2026-01-02 12:00:00",
      "2026-01-02 14:00:00",
      "2026-01-03 16:53:42"
    ),
    tz = "Africa/Lagos"
  ),
  count = c(1, 0, 2, 1, 2)
)
deployment_data <- dplyr::tibble(
  cam = c(1, 2, 2, 2),
  start = as.POSIXct(
    c(
      "2025-12-01 15:00:00",
      "2025-12-08 00:00:00",
      "2026-01-01 00:00:00",
      "2026-01-02 00:00:00"
    ),
    tz = "Africa/Lagos"
  ),
  end = as.POSIXct(
    c(
      "2026-01-05 00:00:00",
      "2025-12-19 03:30:00",
      "2026-01-01 05:00:00",
      "2026-01-05 00:00:00"
    ),
    tz = "Africa/Lagos"
  ),
  area = c(300, 200, 200, 450)
)
ct_fit_ise(data, deployment_data,
       sampling_frequency = 3600,
       sampling_length = 10,
       study_area = 1e6)

}
\seealso{
\code{\link[=ct_fit_tte]{ct_fit_tte()}}, \code{\link[=ct_fit_ste]{ct_fit_ste()}}
}
