% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REM.R
\name{ct_fit_speedmodel}
\alias{ct_fit_speedmodel}
\title{Fit animal speed model}
\usage{
ct_fit_speedmodel(
  formula = speed ~ 1,
  data,
  newdata = NULL,
  distance_unit = c("m", "km", "cm"),
  time_unit = c("second", "minute", "hour", "day"),
  ...
)
}
\arguments{
\item{formula}{A formula indicating how speed should be modeled (e.g., \code{speed ~ 1}).}

\item{data}{A data frame containing speed observations.}

\item{newdata}{Optional new data to use for prediction.}

\item{distance_unit}{Unit of distance. One of \code{"m"}, \code{"km"}, \code{"cm"}.}

\item{time_unit}{Unit of time. One of \code{"second"}, \code{"minute"}, \code{"hour"}, \code{"day"}.}

\item{...}{Additional arguments passed to \code{\link[sbd:sbm]{sbd::sbm()}}.}
}
\value{
An object of class \code{sbm}, with an additional \code{unit} attribute indicating the speed unit.
}
\description{
Fits a statistical model to estimate average movement speed of animals.
Used in the REM density estimation.
}
\examples{
data("ctdp")
observations <- ctdp$data$observations \%>\%
  dplyr::filter(scientificName == "Vulpes vulpes")

ct_fit_speedmodel(speed ~ 1, data = observations)

}
\seealso{
\code{\link[=ct_fit_rem]{ct_fit_rem()}}, \code{\link[=ct_fit_detmodel]{ct_fit_detmodel()}}, \code{\link[=ct_fit_activity]{ct_fit_activity()}}
}
