% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_inext.R
\name{ct_inext}
\alias{ct_inext}
\title{Interpolation and extrapolation of Hill number}
\usage{
ct_inext(
  data,
  species_column,
  site_column,
  size_column,
  strata_column = NULL,
  diversity_order = 0,
  sample_size = NULL,
  endpoint = NULL,
  knots = 40,
  n_bootstrap = 100
)
}
\arguments{
\item{data}{A data frame, preferably the output of \code{\link[=ct_camera_day]{ct_camera_day()}}.}

\item{species_column}{The column in the data frame representing species identifiers. Can be specified as a string or unquoted column name.}

\item{site_column}{The column in the data frame representing site identifiers. Can be specified as a string or unquoted column name.}

\item{size_column}{(Optional) The column representing the size or abundance of the species at each site. If not provided, counts of species occurrences are calculated.}

\item{strata_column}{Optional column name for a grouping variable (e.g. habitat, treatment).
If provided, estimates are computed separately for each stratum.}

\item{diversity_order}{Numeric specifying the order of diversity (q) for Hill numbers.
Common values:
\itemize{
\item \code{0} = species richness,
\item \code{1} = Shannon diversity (exponential of Shannon entropy),
\item \code{2} = Simpson diversity (inverse of Simpson index).
Defaults to \code{0}.
}}

\item{sample_size}{Optional numeric vector specifying sample sizes for interpolation/extrapolation.}

\item{endpoint}{Optional numeric specifying the maximum sample size for extrapolation.
If NULL, endpoint is the double of the current sample size}

\item{knots}{Integer specifying the number of equally spaced knots for rarefaction/extrapolation.
Default is \code{40}.}

\item{n_bootstrap}{Number of bootstrap replications for estimating confidence intervals.
Default is \code{100}.}
}
\value{
A list containing:
\itemize{
\item \code{DataInfo}: Summary statistics of input data.
\item \code{iNextEst}: Rarefaction/extrapolation results for specified diversity order.
\item \code{AsyEst}: Asymptotic diversity estimates.
}
}
\description{
Computes incidence-based species diversity estimates (Hill numbers) from camera trap data.
This is a wrapper around iNEXT package (Chao et al., 2014; Hsieh, Ma, & Chao, 2016).
}
\details{
This function converts the input data into an \strong{incidence-frequency vector}.
The first element of the vector is the number of sampling units, followed by species frequencies.
If \code{strata_column} is provided, the conversion is done separately for each stratum.
}
\examples{
if (requireNamespace("iNEXT", quietly = TRUE)) {
## Import example data
camdata1 <- read.csv(ct:::table_files()[1]) \%>\%
  dplyr::mutate(site = "pene") \%>\%
  # remove consecutive entry of the same species at the same location within 60s
  ct_independence(species_column = species,
                  site_column = camera,
                  datetime = datetimes,
                  threshold = 60, format = "\%Y-\%m-\%d \%H:\%M:\%S"
                  )
head(camdata1)

# Prepare sampling data (camera-day)
camday <- ct_camera_day(
  data = camdata1,
  deployment_column = camera,
  datetime_column = datetime,
  species_column = species,
  size_column = number
)

# RAREFACTION/EXTRAPOLATION
int_ext <- ct_inext(data = camday,
                    diversity_order = c(0, 1, 2),
                    species_column = species,
                    site_column = sampling_unit,
                    size_column = number,
                    n_bootstrap = 50)
int_ext

# plot with curves colored by order
ct_plot_inext(int_ext, type = 1, color_var = "Order.q")

# plot with curves faceted by order
ct_plot_inext(int_ext, type = 1, facet_var = "Order.q")
 }
}
\references{
Chao, A., Gotelli, N. J., Hsieh, T. C., Sander, E. L., Ma, K. H., Colwell,
R. K., & Ellison, A. M. (2014). Rarefaction and extrapolation with
Hill numbers: a framework for sampling and estimation in species diversity
studies. Ecological Monographs, 84, 45–67.
\doi{https://doi.org/https://doi.org/10.1890/13-0133.1}

Hsieh, T. C., Ma, K. H., & Chao, A. (2016). iNEXT: an R package for
rarefaction and extrapolation of species diversity (Hill numbers).
Methods in Ecology and Evolution, 7(12), 1451–1456.
\doi{https://doi.org/10.1111/2041-210X.12613}
}
\seealso{
\code{\link[=ct_camera_day]{ct_camera_day()}} for preparing sampling data (camera-day).
}
