% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXIFTOOL.R
\name{ct_install_exiftool}
\alias{ct_install_exiftool}
\title{Install ExifTool, downloading (by default) the current version}
\usage{
ct_install_exiftool(
  install_location = NULL,
  win_exe = NULL,
  local_exiftool = NULL,
  quiet = FALSE
)
}
\arguments{
\item{install_location}{Path to the directory into which ExifTool should be
installed. If \code{NULL} (the default), installation will be into the
directory returned by \code{ct:::R_user_dir("ct")}.}

\item{win_exe}{Logical, only used on Windows machines. Should we install the
standalone ExifTool Windows executable or the ExifTool Perl library?
(The latter relies, for its execution, on an existing installation of
Perl being present on the user's machine.)  If set to \code{NULL} (the
default), the function installs the Windows executable on Windows
machines and the Perl library on other operating systems.}

\item{local_exiftool}{If installing ExifTool from a local "*.zip" or
".tar.gz", supply the path to that file as a character string. With
default value, \code{NULL}, the function downloads ExifTool from
\url{https://exiftool.org} and then installs it.}

\item{quiet}{Logical.  Should function should be chatty?}
}
\value{
Called for its side effect
}
\description{
Install the current version of ExifTool
}
