% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot_inext.R
\name{ct_plot_inext}
\alias{ct_plot_inext}
\title{Plot diversity interploation and extrapolation}
\usage{
ct_plot_inext(
  inext_object,
  type = 1,
  se = TRUE,
  facet_var = "None",
  color_var = "Assemblage",
  grey = FALSE
)
}
\arguments{
\item{inext_object}{an object as outputed by \code{\link[=ct_inext]{ct_inext()}}}

\item{type}{three types of plots:
\itemize{
\item \code{type = 1}: sample-size-based rarefaction/extrapolation curve
\item \code{type = 2}: sample completeness curve
\item \code{type = 3}: coverage-based rarefaction/extrapolation curve
}}

\item{se}{a logical variable to display confidence interval
around the estimated sampling curve.}

\item{facet_var}{create a separate plot for each value of a specified variable:
\itemize{
\item \code{facet_var = "None"}: no separation
\item \code{facet_var = "Order.q"}: a separate plot for each diversity order
\item \code{facet_var = "Assemblage"}: a separate plot for each assemblage
\item \code{facet_var = "Both"}: a separate plot for each combination of order x assemblage
}}

\item{color_var}{create curves in different colors for values of a
specified variable:
\itemize{
\item \code{color_var = "None"}: all curves are in the same color
\item \code{color_var = "Order.q"}: use different colors for diversity orders
\item \code{color_var = "Assemblage"}: use different colors for sites
\item \code{color_var = "Both"}: use different colors for combinations of order x assemblage
}}

\item{grey}{a logical variable to display grey and white ggplot2 theme}
}
\value{
a ggplot2 object
}
\description{
plot sample-size-based and coverage-based rarefaction/extrapolation curves along
with a bridging sample completeness curve
}
\examples{
if (requireNamespace("iNEXT", quietly = TRUE)) {
## Import example data
camdata1 <- read.csv(ct:::table_files()[1]) \%>\%
  dplyr::mutate(site = "pene") \%>\%
  # remove consecutive entry of the same species at the same location within 60s
  ct_independence(species_column = species,
                  site_column = camera,
                  datetime = datetimes,
                  threshold = 60, format = "\%Y-\%m-\%d \%H:\%M:\%S"
                  )
head(camdata1)

# Prepare sampling data (camera-day)
camday <- ct_camera_day(
  data = camdata1,
  deployment_column = camera,
  datetime_column = datetime,
  species_column = species,
  size_column = number
)

# RAREFACTION/EXTRAPOLATION
int_ext <- ct_inext(data = camday,
                    diversity_order = c(0, 1, 2),
                    species_column = species,
                    site_column = sampling_unit,
                    size_column = number,
                    n_bootstrap = 50)
int_ext

# plot with curves colored by order
ct_plot_inext(int_ext, type = 1, color_var = "Order.q")

# plot with curves faceted by order
ct_plot_inext(int_ext, type = 1, facet_var = "Order.q")
 }
}
