% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_to_time.R
\name{ct_to_time}
\alias{ct_to_time}
\title{Convert radian to time}
\usage{
ct_to_time(radian)
}
\arguments{
\item{radian}{A numeric value or vector representing an angle in radians.
The value must lie within the range \eqn{[0, 2\pi]}, where 0 corresponds to midnight
(00:00:00) and \eqn{2\pi} corresponds to the next midnight (24:00:00).}
}
\value{
A character string representing the time in the format '\%H:\%M:\%S'.
}
\description{
This function converts an angle in radians (representing a fraction of a full circle)
into a time in the format '\%H:\%M:\%S'. The conversion assumes that the radian value
represents a fraction of a 24-hour day (i.e., 0 radians is midnight
and \eqn{2\pi} radians is the next midnight).
}
\examples{
# Convert 1.6 radians to time
ct_to_time(1.6)
# Output: "06:06:42"

}
\seealso{
\code{\link[=ct_to_radian]{ct_to_radian()}}
}
