% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traprate.R
\name{ct_traprate_estimate}
\alias{ct_traprate_estimate}
\title{Estimate trap rate}
\usage{
ct_traprate_estimate(data, strata = NULL, n_bootstrap = 1000)
}
\arguments{
\item{data}{A data frame as returned by \code{\link[=ct_traprate_data]{ct_traprate_data()}} with columns \code{n} and \code{effort}.}

\item{strata}{Optional. A data frame defining strata, with columns \code{stratumID} and \code{area}.}

\item{n_bootstrap}{Number of bootstrap replicates to estimate uncertainty. Default is 1000.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{estimate}: Trap rate estimate (e.g., detections per day)
\item \code{se}: Standard error of the estimate
\item \code{cv}: Coefficient of variation
\item \code{lower_ci}: Lower bound of the 95\\% confidence interval
\item \code{upper_ci}: Upper bound of the 95\\% confidence interval
\item \code{n}: Number of deployments or observation used
\item \code{unit}: Effort unit
}
}
\description{
Computes the estimated trap rate and uncertainty using bootstrapping, with optional
support for stratified estimation based on area-weighted averaging.
}
\examples{

data("ctdp")
deployments <- ctdp$data$deployments
observations <- ctdp$data$observations \%>\%
                  dplyr::filter(scientificName == "Vulpes vulpes")

trap_rate <- ct_traprate_data(observation_data = observations,
                              deployment_data = deployments,
                              use_deployment = FALSE,
                              deployment_column = deploymentID,
                              datetime_column = timestamp,
                              start = start, end = 'end'
)

ct_traprate_estimate(data = trap_rate, n_bootstrap = 1000)

}
\seealso{
\code{\link[=ct_get_effort]{ct_get_effort()}}, \code{\link[=ct_traprate_data]{ct_traprate_data()}}
}
