% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXIFTOOL.R
\name{download_exiftool}
\alias{download_exiftool}
\title{Download the current version of ExifTool}
\usage{
download_exiftool(win_exe = FALSE, download_path = NULL, quiet = FALSE)
}
\arguments{
\item{win_exe}{Logical, only used on Windows machines. Should we install the
standalone ExifTool Windows executable or the ExifTool Perl library?
(The latter relies, for its execution, on an existing installation of
Perl being present on the user's machine.)  If set to \code{NULL} (the
default), the function installs the Windows executable on Windows
machines and the Perl library on other operating systems.}

\item{download_path}{Path indicating the location to which
ExifTool should be downloaded.}

\item{quiet}{Logical.  Should function should be chatty?}
}
\value{
A character string giving the path to the downloaded
ExifTool.
}
\description{
Download the current version of ExifTool
}
\author{
Joshua O'Brien
}
\keyword{internal}
