% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTDS.R
\docType{data}
\name{duikers}
\alias{duikers}
\title{Maxwell's duiker camera-trap distance & video-start data}
\format{
A named list with these tibbles:
\strong{DaytimeDistances}: A tibble of all Maxwell's duiker distance
observations (including non peak periods) recorded at camera stations
during daytime deployments. It has the following columns:
\itemize{
\item \code{distance}: the midpoint (m) of the assigned distance interval between animal and camera.
\item \code{Sample.Label}: camera station identifier.
\item \code{Effort}: number of active 2 second time steps the camera operated (i.e. temporal effort).
\item \code{Region.Label}: stratum name (only a single stratum in this dataset).
\item \code{Area}: study area size (km^2; in this dataset, 40.4).
\item \code{multiplier}: spatial effort: fraction of a full circle covered, based on
the camera's 42 deg field of view (42/360).
\item \code{utm.e}: UTM easting (metres) of the camera station.
\item \code{utm.n}: UTM northing (metres) of the camera station.
\item \code{object}: a unique identifier for each observation
}

\strong{PeakDistances}: A tibble with the \strong{same column structure} as
\strong{DaytimeDistances}, but includes \strong{only} observations during the
species' peak activity periods (no dawn or late day records).

\strong{VideoStartTimesFullDays}: A tibble of camera‑trigger times for duiker
videos that were recorded on \strong{full day deployments} (i.e. days without
researcher visits). Columns include:
\itemize{
\item \code{order}: sequential order of video events at each
station/day.
\item \code{folder}: local folder (e.g. \code{"A1"}) for grouping
videos by station or session.
\item \code{vid.no}: unique video identifier number.
\item \code{ek.no}: event key number (original trigger event id).
\item \code{easting}: UTM easting (metres) for the video event.
\item \code{northing}: UTM northing (metres) for the video event.
\item \code{month}: calendar month (1-12) of the video event.
\item \code{day}: day of the month (1-31).
\item \code{hour}: hour (24h clock) when the video started.
\item \code{minute}: minute of that hour when the video started.
\item \code{date}: Date of the record
\item \code{time}: Time of the record
\item \code{datetime}: date pasted with time
}
}
\usage{
duikers
}
\description{
The \strong{\code{duikers}} dataset is a \strong{named list} of three tibbles derived from
Maxwell's duiker (\emph{Philantomba maxwellii}) camera trap and distance sampling
data collected in Taï National Park, Côte d'Ivoire (2014), and archived as
the Dryad dataset \emph{Distance sampling with camera traps} (Howe et al., 2018)
}
\references{
Howe, E. J., Buckland, S. T., Després-Einspenner, M. L., Kühl, H. S., & Buckland, S. T. (2018).
Data from: Distance sampling with camera traps. \doi{10.5061/dryad.b4c70}
}
\keyword{datasets}
