% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXIFTOOL.R
\name{exiftool_version}
\alias{exiftool_version}
\title{Get ExifTool Version}
\usage{
exiftool_version(exiftool_path = NULL)
}
\arguments{
\item{exiftool_path}{Path to ExifTool executable (auto-detected if NULL)}
}
\value{
Character string with version number
}
\description{
Check the installed version of ExifTool
}
\keyword{internal}
