% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXIFTOOL.R
\name{find_exiftool}
\alias{find_exiftool}
\title{Find ExifTool Executable}
\usage{
find_exiftool(install_location = NULL)
}
\arguments{
\item{install_location}{Optional custom installation location to check first}
}
\value{
Path to ExifTool executable or NULL if not found
}
\description{
Locates the ExifTool executable on the system
}
\keyword{internal}
