% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_get_hs.R
\name{parse_hs}
\alias{parse_hs}
\title{Parse Hierarchical Subjects into a Tibble}
\usage{
parse_hs(hs_vector, hs_delimitor = "|")
}
\description{
Parse Hierarchical Subjects into a Tibble
}
\keyword{internal}
