% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{R_user_dir}
\alias{R_user_dir}
\title{Determine the Per-User Directory for R Package Data, Config, or Cache}
\usage{
R_user_dir(package, which = c("data", "config", "cache"))
}
\arguments{
\item{package}{Character string giving the package name.}

\item{which}{Character string specifying the directory type.
Must be one of \code{"data"}, \code{"config"}, or \code{"cache"}.}
}
\value{
A character string giving the full path to the package-specific
per-user directory.
}
\description{
Determine the per-user directory where packages can store data,
configuration files, or caches.
}
\keyword{internal}
