% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-package.R
\docType{package}
\name{ct-package}
\alias{ct}
\alias{ct-package}
\title{ct: Integrated Camera-Trap Data Management and Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An integrated, tidyverse-friendly workflow for camera trap data in wildlife monitoring and ecological research. Reads and edits media metadata, filters independent detections, analyses activity patterns and species diversity, and estimates species density or abundance with several methods, including the random encounter model, camera-trap distance sampling, time-to-event, space-to-event, and the random encounter and staying-time model (see Rowcliffe et al. (2008) \doi{10.1111/j.1365-2664.2008.01473.x}, Howe et al. (2017) \doi{10.1111/2041-210X.12790}, Nakashima et al. (2018) \doi{10.1111/1365-2664.13059}, and Moeller et al. (2018) \doi{10.1002/ecs2.2331}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stangandaho.github.io/ct/}
  \item Report bugs at \url{https://github.com/stangandaho/ct/issues}
}

}
\author{
\strong{Maintainer}: Stanislas Mahussi Gandaho \email{stangandaho@gmail.com} (\href{https://orcid.org/0009-0003-4969-1252}{ORCID})

Other contributors:
\itemize{
  \item Pablo Palencia \email{palencia.pablo.m@gmail.com} (\href{https://orcid.org/0000-0002-2928-4241}{ORCID}) [contributor, reviewer]
  \item R Consortium [funder]
}

}
\keyword{internal}
