% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-package.R
\docType{package}
\name{ct-package}
\alias{ct}
\alias{ct-package}
\title{ct: Integrated Camera-Trap Data Management and Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An integrated, tidyverse-friendly workflow for camera trap data in wildlife monitoring and ecological research. Reads and edits image metadata, filters independent detections, analyses activity patterns and species diversity, prepares occupancy-format data, and estimates species density or abundance with several methods, including the random encounter model, camera-trap distance sampling, time-to-event, space-to-event, and the random encounter and staying-time model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stangandaho.github.io/ct/}
  \item Report bugs at \url{https://github.com/stangandaho/ct/issues}
}

}
\author{
\strong{Maintainer}: Stanislas Mahussi Gandaho \email{stangandaho@gmail.com} (\href{https://orcid.org/0009-0003-4969-1252}{ORCID})

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
\keyword{internal}
