% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-package.R
\docType{package}
\name{ct-package}
\alias{ct}
\alias{ct-package}
\title{ct: Camera Trap Data Tool}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tool to edit, extract, visualize, and analyze camera trap data for wildlife monitoring and ecological research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stangandaho.github.io/ct/}
  \item Report bugs at \url{https://github.com/stangandaho/ct/issues}
}

}
\author{
\strong{Maintainer}: Stanislas Mahussi Gandaho \email{stangandaho@gmail.com} (\href{https://orcid.org/0009-0003-4969-1252}{ORCID})

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
\keyword{internal}
