% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_boot_ci.R
\name{ct_boot_ci}
\alias{ct_boot_ci}
\title{Bootstrap confidence intervals}
\usage{
ct_boot_ci(t0, bt, conf = 0.95)
}
\arguments{
\item{t0}{the statistic estimated from the original sample, usually the output from \code{\link[=ct_overlap_estimates]{ct_overlap_estimates()}}}

\item{bt}{a vector of bootstrap statistics, usually the output from \code{\link[=ct_boot_estimates]{ct_boot_estimates()}}}

\item{conf}{
a (single!) confidence interval to estimate.
}
}
\description{
Confidence interval calculation from bootstrap samples.
}
