% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp.R
\name{ct_dp_filter}
\alias{ct_dp_filter}
\title{Filter camera trap data package}
\usage{
ct_dp_filter(package, table = c("observations", "deployments", "media"), ...)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
ct_read_dp().}

\item{table}{Character indicating the table to read - one "observations",
"deployments", or "media"}

\item{...}{Filtering conditions, see dplyr::filter()}
}
\description{
Subsets observations in camera trap data package, retaining all rows that
satisfy the conditions.
}
\examples{
\donttest{
dp <- ct_dp_example()
ct_dp_filter(package = dp, table = "observation",
scientificName == "Vulpes vulpes", observationLevel == "event"
)

ct_dp_filter(package = dp, table = "deployments",
             latitude > 51.0, longitude > 5.0)


ct_dp_filter(package = dp, table = "media",
             captureMethod == "activityDetection", filePublic == FALSE
)
}
}
