% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_get_hs.R
\name{ct_get_hs}
\alias{ct_get_hs}
\title{Retrieve hierarchical subject values from image metadata}
\usage{
ct_get_hs(path)
}
\arguments{
\item{path}{A character vector specifying the full path of the image file.}
}
\value{
A character vector of unique hierarchical subjects if they exist, otherwise NULL.
}
\description{
The function retrieves the values from the \verb{Hierarchical Subject}
field of an image's metadata. It uses exiftool to read the metadata and processes
the results to extract and return the unique hierarchical subjects.
}
\examples{

# Image path
image_path <- file.path(system.file("img", package = "ct"), "large.jpeg")

# Get Hierarchical Subject from the image
ct_get_hs(path = image_path)

}
