% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_remove_hs.R
\name{ct_remove_hs}
\alias{ct_remove_hs}
\title{Remove specific or all hierarchical subject values from image metadata}
\usage{
ct_remove_hs(path, hierarchy = NULL, intern = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{path}{A character vector specifying the full path of the image file.}

\item{hierarchy}{A named character vector, e.g c("Species" = "Vulture") specifying the hierarchy to be removed.
If NULL, the entire \verb{Hierarchical Subject} field is removed.}

\item{intern}{TRUE if output should be returned as a character vector.}

\item{quiet}{Suppress output of the command itself.}

\item{...}{additional arguments to be passed to \code{\link[=system2]{system2()}}}
}
\value{
message indicating image updated
}
\description{
The ct_remove_hs function removes a specific hierarchy from the
\verb{Hierarchical Subject} field in an image's metadata, or it removes the entire
field if no specific hierarchy is provided. It uses exiftool
to manipulate the metadata and ensure that only the desired changes are applied.
}
\examples{

# Image path
image_path <- file.path(system.file("img", package = "ct"), "large.jpeg")

# Get Hierarchical Subject from the image
no_hs <- ct_get_hs(path = image_path)
ct_create_hs(image_path, c("A" = "AB"))
ct_remove_hs(image_path, c("A" = "AB"))

}
